pyproject.toml
[project]
name = "expense-tracker-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0.2",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
]
